requirements/requirements.txt
requests==2.26.0
aiohttp==3.8.1
pydantic==1.9.1
mariadb==1.0.9
sqlalchemy[asyncio]==1.4.31
alembic==1.7.1
mysqlclient==2.1.0
PyMySQL==1.0.2
asyncmy==0.2.3
flask-log-request-id==0.10.1 # for bst logger
core-shared-libs==4.0.0
openapi-spec-validator==0.4.0
connexion[swagger-ui]==2.6.0
python-jose==3.3.0
uvicorn==0.16.0
fastapi==0.78.0
httpx==0.23.0

opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
